mesh-adoption/quality/quality-sidecar/requirements.txt
great_expectations==1.3.5
pandas
opentelemetry-api==1.28.2
opentelemetry-distro==0.49b2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
wrapt==1.16.0
pytest
pyspark
python-dotenv
sh